@@ -4,8 +4,7 @@ description: Boton API
 homepage: homepage
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
-
+  sdk: ">=3.0.6 <4.0.0"
 dependencies:
   dio: "^5.2.0"
   one_of: ">=1.5.0 <2.0.0"
@@ -16,4 +15,4 @@ dependencies:
 dev_dependencies:
   built_value_generator: ">=8.4.0 <9.0.0"
   build_runner: any
-  test: ^1.16.0 # Dependencia para realizar pruebas unitarias
+  test: ^1.16.0
