@@ -51,23 +51,92 @@ class _SettingsViewState extends State<SettingsView> {
     });
   }
 
-  void _updateUserData(String field, String value) {
-    final user = widget.controller.user?.copyWith(
-      nombreCompleto: field == 'nombreCompleto'
-          ? value
-          : widget.controller.user?.nombreCompleto,
-      matricula:
-          field == 'matricula' ? value : widget.controller.user?.matricula,
-      numeroTelefono: field == 'numeroTelefono'
-          ? value
-          : widget.controller.user?.numeroTelefono,
-      correoElectronico: field == 'correoElectronico'
-          ? value
-          : widget.controller.user?.correoElectronico,
-    );
-    if (user != null) {
-      widget.controller.updateUser(user);
+  void _updateUserData() {
+    final nombreCompleto = _nombreCompletoController.text;
+    final matricula = _matriculaController.text;
+    final numeroTelefono = _numeroTelefonoController.text;
+    final correoElectronico = _correoElectronicoController.text;
+    print("numeroTelefono: $numeroTelefono");
+    print("matricula: $matricula");
+    if (_isValidMatricula(matricula) &&
+        _isValidNumeroTelefono(numeroTelefono) &&
+        _isValidCorreoElectronico(correoElectronico)) {
+      final user = widget.controller.user?.copyWith(
+        nombreCompleto: nombreCompleto.isNotEmpty ? nombreCompleto : null,
+        matricula: matricula.isNotEmpty ? matricula : null,
+        numeroTelefono: numeroTelefono.isNotEmpty ? numeroTelefono : null,
+        correoElectronico:
+            correoElectronico.isNotEmpty ? correoElectronico : null,
+      );
+      if (user != null) {
+        widget.controller.updateUser(user);
+        _showModificadoCorrectamente();
+      }
+    } else {
+      _showErrorDialog();
     }
+  }
+
+  _showModificadoCorrectamente() {
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('Datos Modificados'),
+          content: const Text('Los datos se han modificado correctamente.'),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () => Navigator.of(context).pop(),
+              child: const Text('Aceptar'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  bool _isValidMatricula(String matricula) {
+    if (matricula.isEmpty) {
+      return true;
+    }
+    final numericMatricula = int.tryParse(matricula);
+    return numericMatricula != null &&
+        matricula.length >= 10 &&
+        matricula.length <= 12;
+  }
+
+  bool _isValidNumeroTelefono(String numeroTelefono) {
+    if (numeroTelefono.isEmpty) {
+      return true;
+    }
+    final numericTelefono = int.tryParse(numeroTelefono);
+    return numericTelefono != null &&
+        numeroTelefono.length >= 8 &&
+        numeroTelefono.length <= 12;
+  }
+
+  bool _isValidCorreoElectronico(String correoElectronico) {
+    final emailRegExp = RegExp(r'^[^@]+@[^@]+\.[^@]+$');
+    return correoElectronico.isEmpty || emailRegExp.hasMatch(correoElectronico);
+  }
+
+  void _showErrorDialog() {
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('Datos inválidos'),
+          content: const Text(
+              'Por favor, verifica que todos los datos sean correctos.'),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () => Navigator.of(context).pop(),
+              child: const Text('Aceptar'),
+            ),
+          ],
+        );
+      },
+    );
   }
 
   @override
@@ -85,6 +154,13 @@ class _SettingsViewState extends State<SettingsView> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
+                  // Mensaje informativo
+                  const Padding(
+                    padding: EdgeInsets.only(bottom: 16),
+                    child: Text(
+                        'Aquí puedes modificar tus datos. Estos datos personales no serán utilizados solo hasta que envíes un reporte. Nuestros servidores no almacenan tus datos personales en ningún momento, solo se utilizan para identificarte en caso de que envíes un reporte.',
+                        style: TextStyle(fontSize: 16)),
+                  ),
                   // Sección de información del usuario
                   const Text(
                     'Información del Usuario',
@@ -97,18 +173,27 @@ class _SettingsViewState extends State<SettingsView> {
                   ),
                   _buildTextField(
                     controller: _matriculaController,
-                    label: 'Matrícula o numero de empleado',
+                    label: 'Matrícula o Número de Empleado',
                     field: 'matricula',
+                    keyboardType: TextInputType.number,
                   ),
                   _buildTextField(
                     controller: _numeroTelefonoController,
                     label: 'Número de Teléfono',
                     field: 'numeroTelefono',
+                    keyboardType: TextInputType.number,
                   ),
                   _buildTextField(
                     controller: _correoElectronicoController,
                     label: 'Correo Electrónico',
                     field: 'correoElectronico',
+                    keyboardType: TextInputType.emailAddress,
+                  ),
+                  const SizedBox(height: 20),
+                  // Botón para guardar los datos
+                  ElevatedButton(
+                    onPressed: _updateUserData,
+                    child: const Text('Guardar'),
                   ),
                   const SizedBox(height: 20),
                   // Sección para la selección del tema
@@ -148,6 +233,7 @@ class _SettingsViewState extends State<SettingsView> {
     required TextEditingController controller,
     required String label,
     required String field,
+    TextInputType? keyboardType,
   }) {
     return Padding(
       padding: const EdgeInsets.only(bottom: 16),
@@ -155,17 +241,8 @@ class _SettingsViewState extends State<SettingsView> {
         controller: controller,
         decoration: InputDecoration(
           labelText: label,
-          suffixIcon: IconButton(
-            icon: const Icon(Icons.edit),
-            onPressed: () {
-              _updateUserData(field, controller.text);
-            },
-          ),
         ),
-        onChanged: (value) {
-          // Actualizar los datos del usuario en tiempo real
-          _updateUserData(field, value);
-        },
+        keyboardType: keyboardType,
       ),
     );
   }
